All nodes set the node's hash with StreamDataImpl.set_hash().
This fixes incorrect evaluations of the graph when a single node
was connected to two different output (write image) nodes. In this
case, when evaluating the second write image node, the node would
incorrectly use the previously computed (cached) value because the
hashes collided. This is now fixed.

The problem could be found in the C++ test_f example with the
"execute2" section.