Refactor PixelBlock, into BlockSize and DataBlock. Renamed PixelDataType.
This is an API breaking change.

PixelDataType was renamed to DataType, as this enum can be re-used.

The data type conversions have completely changed into an enum.
This allows us to reduce the amount of pretending the same f32 memory
is actually u8 and reduces the amount of size conversion bytes needed.