Ensure cache is enabled by default, to stop bugs from happening.
There are currently bugs that can happen when the cache is turned
off for some nodes. For example if a viewer node is connected to a
read node and the viewer node is set to bake the input, and the
read node has use_cache=0, then the viewer node will be forced to
re-read the image file each time, even if the pixel data it really
only needs is inside the viewer node cache.